Fix single module with Native Build Tools use case
If the native profile only applies when the plugin is defined, this does
not really help for the single module use case. A single module should
declare the spring boot plugin.

Note that, although the native build tools plugin has not been applied,
the `compile` goal fetches the metadata if necessary.