Fix multi modules with Buildpacks use case
We can't use spring-boot:build-image in a multi modules project as it
will attempt to build an image for every module. Rather we should opt-in
for it in the application module(s).

This commit uses a profile named 'native' so that the behavior of the
parent is reused with the same name.