Fix multi modules with Native Build Tools use case
We can't use native:compile in a multi modules project as it will
attempt to build an image for every module. Rather we should opt-in
for it in the application module(s).

This commit uses a profile named 'native' so that the behavior of the
parent is reused with the same name. Note that we use the "no-fork"
goal so that it doesn't run package again. This is not available yet
for our Maven Plugin.